Tidy repo worktree frontend maintainability
